% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinystring-package.R
\docType{package}
\name{tinystring-package}
\alias{tinystring}
\alias{tinystring-package}
\title{tinystring: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Laura Bakala \email{erdaradun.gaztea@gmail.com} (\href{https://orcid.org/0000-0002-3213-2484}{ORCID}) [copyright holder]

}
\keyword{internal}
