% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c.R
\name{tstr_c}
\alias{tstr_c}
\title{Join multiple strings}
\usage{
tstr_c(..., sep = character())
}
\arguments{
\item{...}{[\code{tstr()}]\cr
TinyStrings to join.}

\item{sep}{[\code{character()}]\cr
Separator to insert between strings; each element of the vector is treated as a separate letter. \code{NA} characters
are allowed.}
}
\value{
TinyString with as many elements as \link[vctrs::vec_size]{the common size} of the input.
}
\description{
TODO
}
