% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten.R
\name{tstr_flatten}
\alias{tstr_flatten}
\title{Flatten a string}
\usage{
tstr_flatten(x, collapse = character())
}
\arguments{
\item{x}{[\code{tstr()}]\cr
TinyString to modify.}

\item{collapse}{[\code{character()}]\cr
Separator to insert between strings; each element of the vector is treated as a separate letter. \code{NA} characters
are allowed.}
}
\value{
TinyString with one element.
}
\description{
TODO
}
\examples{
tstr_pack(c("rllssllsrrlr", "slllsr"), alphabet = c("s", "l", "r")) |>
  tstr_flatten() |>
  tstr_unpack()

}
