% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather.R
\name{tstr_gather}
\alias{tstr_gather}
\title{Collapse letters into manually defined groups}
\usage{
tstr_gather(x, ..., .unmentioned = NULL)
}
\arguments{
\item{x}{[\code{tstr()}]\cr
TinyString to modify.}

\item{...}{[\code{character()}]\cr
Groups to collapse. All elements of each vector are replaced with parameter name.}

\item{.unmentioned}{[\code{character(1)}]\cr
What to do with letters not listed in \code{...}. If \code{NULL}, they are kept as-is. If \code{NA}, they are dropped and replaced
with missing letter. If a single string, they are all replaced with this string.}
}
\value{
Modified TinyString.
}
\description{
TODO
}
