% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pack.R
\name{tstr_pack}
\alias{tstr_pack}
\title{Create compressed string representation}
\usage{
tstr_pack(x, alphabet = letters)
}
\arguments{
\item{x}{[\code{character()}]\cr
A character vector to compress.}

\item{alphabet}{[\code{character()}]\cr
Set of "letters" to use when compressing \code{x}.}
}
\value{
A TinyString object the same length as \code{x}.
}
\description{
TODO
}
