% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{tstr_recode}
\alias{tstr_recode}
\title{Recode letters}
\usage{
tstr_recode(x, ...)
}
\arguments{
\item{x}{[\code{tstr()}]\cr
TinyString to modify.}

\item{...}{[\code{character(1)}]\cr
Letters to recode; parameter names define target letters.}
}
\description{
TODO
}
\examples{
tstr_pack(c("rllssllsrr", "slllsrssl"), alphabet = c("s", "l", "r")) |>
  tstr_recode("u" = "s") |>
  tstr_unpack()

}
