% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub.R
\name{tstr_sub}
\alias{tstr_sub}
\title{Get substrings using their positions}
\usage{
tstr_sub(x, start = 1L, end = min(tstr_length(x)))
}
\arguments{
\item{x}{[\code{tstr()}]\cr
A TinyString object.}

\item{start, end}{[\code{integer(1)}]\cr
Indices defining the range of characters to extract (inclusive).}
}
\value{
A TinyString object the same length as \code{x}.
}
\description{
TODO
}
\examples{
tstr_pack(c("rllssllsrrlr", "slllsr"), alphabet = c("s", "l", "r")) |>
  tstr_sub(3, 6)

}
