% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pack.R
\name{tstr_unpack}
\alias{tstr_unpack}
\title{Decompress a TinyString object}
\usage{
tstr_unpack(x)
}
\arguments{
\item{x}{[\code{tstr()}]\cr
A TinyString object.}
}
\value{
A character vector the same length as \code{x}.
}
\description{
TODO
}
