% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length.R
\name{length.tstr}
\alias{length.tstr}
\alias{length<-.tstr}
\title{Number of TinyStrings}
\usage{
\method{length}{tstr}(x)

\method{length}{tstr}(x) <- value
}
\arguments{
\item{x}{[\code{tstr()}]\cr
A TinyString object.}

\item{value}{[\code{integer(1)}]\cr
For the setter function, new length for the \code{x} object.}
}
\value{
A single integer with the number of strings.
}
\description{
TODO
}
