% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinystring-package.R
\docType{package}
\name{tinystring-package}
\alias{tinystring}
\alias{tinystring-package}
\title{tinystring: Optimized Storage of Strings with Limited Character Sets}
\description{
Allows the user to efficiently store and process strings with limited character sets, reducing memory per character to sub-byte values.
}
\author{
\strong{Maintainer}: Laura Bakala \email{erdaradun.gaztea@gmail.com} (\href{https://orcid.org/0000-0002-3213-2484}{ORCID}) [copyright holder]

}
\keyword{internal}
