% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length-string.R
\name{tstr_length}
\alias{tstr_length}
\title{Compute the length}
\usage{
tstr_length(x)
}
\arguments{
\item{x}{[\code{tstr()}]\cr
A TinyString object.}
}
\value{
A numeric vector the same length as \code{x}.
}
\description{
TODO
}
\examples{
tstr_pack(c("rllssllsrrlr", "slllsr"), alphabet = c("s", "l", "r")) |>
  tstr_length()

}
